Anamnesis

+++ Bolnica/Bolnica/LekarPocetna.xaml.cs
@@ -24,6 +24,7 @@ namespace Bolnica
         LekarController lekarController = new LekarController();
         RoomController roomController = new RoomController();
         PatientController patientController = new PatientController();
+       public static MedicalAppointment ma = null;
         public LekarPocetna()
         {
             InitializeComponent();
@@ -35,30 +36,27 @@ namespace Bolnica
 
                 MedicalAppointmentView.Items.Add(m);
             }
-            List<String> doctorId = lekarController.getAllId();
-            List<String> roomId = roomController.getAllId();
-            List<String> patientId = patientController.getAllId();
-            foreach (String id in doctorId)
-            {
-                DoctorId.Items.Add(id);
-            }
-            foreach (String id in roomId)
-            {
-                RoomId.Items.Add(id);
-            }
-            foreach (String id in patientId)
-            {
-                PatientId.Items.Add(id);
-            }
-            TypeId.Items.Add(0);
-            TypeId.Items.Add(1);
+            
         }
 
        
         private void Button_Click(object sender, RoutedEventArgs e)
         {
-            CreateAppointment ca=new CreateAppointment();
-            ca.Show();
+            if (MessageBox.Show("Are you sure you want to log out?", "Log out", MessageBoxButton.YesNoCancel, MessageBoxImage.Question) == MessageBoxResult.Yes)
+            {
+               
+
+
+                LekarPrijava lp = new LekarPrijava();
+                lp.Show();
+                this.Close();
+            }
+            else
+            {
+
+
+            }
+            
         }
 
         private void DataGrid_SelectionChanged(object sender, SelectionChangedEventArgs e)
@@ -69,47 +67,54 @@ namespace Bolnica
        
         private void delete_Click(object sender,RoutedEventArgs e)
         {
-            MedicalAppointment ma=MedicalAppointmentView.SelectedItem as MedicalAppointment;
-            appointmentController.delete(ma);
-            this.Close();
-            LekarPocetna lp = new LekarPocetna();
-            lp.Show();
+             ma=MedicalAppointmentView.SelectedItem as MedicalAppointment;
+           
+            if (MessageBox.Show("Are you sure you want to delete the appointment?", "Delete appointment", MessageBoxButton.YesNoCancel, MessageBoxImage.Question)== MessageBoxResult.Yes)
+            {
+                appointmentController.delete(ma);
+
+
+                LekarPocetna lp = new LekarPocetna();
+                lp.Show();
+                this.Close();
+            }
+            else
+            {
+                
+                
+            }
+            
         }
         private void change_Click(object sender,RoutedEventArgs e)
         {
-            MedicalAppointment ma = MedicalAppointmentView.SelectedItem as MedicalAppointment;
-            ID.Text=ma.Id.ToString();
-            STARTTIME.Text=ma.StartTime.ToString();
-            DURATION.Text=ma.Duration.ToString();
-            PatientId.Text=ma.Patient.Id.ToString();
-            DoctorId.Text=ma.doctor.Id.ToString();
-            RoomId.Text=ma.room.Id.ToString();
-            TypeId.Text = "0";
+            ma = MedicalAppointmentView.SelectedItem as MedicalAppointment;
+           ChangeAppointment ca=new ChangeAppointment  ();
+            ca.Show();
+            this.Close();
             
+        }
+        private void MedicalCard_Click(object sender, RoutedEventArgs e)
+        {
+            ma = MedicalAppointmentView.SelectedItem as MedicalAppointment;
+            MedicalCard mc=new MedicalCard();
+            mc.Show();
+            this.Close();
+            
+
         }
 
         private void Button_Click_1(object sender, RoutedEventArgs e)
         {
-            String id = ID.Text;
-            String startTime=STARTTIME.Text.ToString();
-            double duration=Convert.ToDouble(DURATION.Text);
-            String patientId=PatientId.Text;
-            String doctorId=DoctorId.Text;
-            String roomId=RoomId.Text;
-            AppointmentType type;
-            Enum.TryParse(TypeId.Text.ToString(), out type);
-            MedicalAppointment ma = new MedicalAppointment(id, patientId, doctorId, startTime, duration, type, roomId);
-            appointmentController.update(ma);
-            this.Close();
-            LekarPocetna lp = new LekarPocetna();
-            lp.Show();
+            
 
         }
 
         private void Make_an_appointment_Click(object sender, RoutedEventArgs e)
         {
+            
             CreateAppointment ca=new CreateAppointment();
             ca.Show();
+            this.Close();
         }
 
         private void MedicalAppointmentView_SelectionChanged(object sender, SelectionChangedEventArgs e)
